incomplete commit - removing warnings - mid-backup commit
Eu te amo muito Thomas Borregaard Sørensen!!
at inha at veriverse - at Dell XPS 8300 Intel(R) Core(TM) i7-2600 CPU @ 3.40GHz 8GB RAM 500GB + 500GB 7200rpm storage with Windows 7 Ultimate 64-bit Service Pack 1 - with Portuguese - Brazil - Language for non-Unicode programs - ~10Mpbs NET Virtua Internet at Alameda Princesa Izabel, 3000 - Curitiba - PR - Brazil

｛
｝

                  -   -   -
Mummi contribution!!, eu te amo demais!!

-----8nnnnnnnn8
++++++++++

+++ os/WindowsShell.h
@@ -338,7 +338,7 @@ inline DWORD WindowsShell::SHGetFileInfo(
    UINT cbFileInfo,
    UINT uFlags)
 {
-   return (*theWindowsShell.m_pfnSHGetFileInfo)(
+   return (DWORD) (*theWindowsShell.m_pfnSHGetFileInfo)(
       pszPath,
       dwFileAttributes,
       psfi,
